continue work on the recycler view that will display list of comics, issues with the network call

+++ app/src/main/java/com/example/android/myapplication/model/CurrentXkcdComic.java
@@ -87,7 +87,7 @@ public class CurrentXkcdComic implements Parcelable {
         return month;
     }
 
-    public long getNum() {
+    public Integer getNum() {
         return num;
     }
